requirements.txt
# Core Framework
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
flask-limiter==3.5.0

# API Clients
openai==1.97.1  # OpenAI Python SDK
exa-py==1.14.18   # Exa API Python SDK
tiktoken==0.7.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
redis==5.0.1

# Configuration
PyYAML==6.0.1

# Async Support
aiohttp==3.9.1
asyncio==3.4.3

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Environment & Config
python-dotenv==1.0.0
pydantic==2.11.7

# Utilities
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0

# YouTube Download
yt-dlp==2025.7.21

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Background Tasks
celery==5.3.4
apscheduler==3.10.4

# Security
cryptography==41.0.7
pyjwt==2.8.0

# HTTP Client
httpx==0.25.2
requests==2.31.0